+++ e2e-tests/production-runtime/gatsby-browser.js
@@ -1,3 +1,5 @@
+require(`./src/index.css`)
+
 if (typeof window !== `undefined`) {
   window.___PageComponentLifecycleCallsLog = []
 }