+++ packages/gatsby-source-shopify/src/lib.js
@@ -47,9 +47,9 @@ export const queryAll = async (
   const edges = get([...path, `edges`], data)
   const nodes = edges.map(edge => edge.node)
 
-  aggregatedResponse
-    ? (aggregatedResponse = aggregatedResponse.concat(nodes))
-    : (aggregatedResponse = nodes)
+  aggregatedResponse = aggregatedResponse
+    ? aggregatedResponse.concat(nodes)
+    : nodes
 
   if (get([...path, `pageInfo`, `hasNextPage`], false, data))
     return queryAll(