+++ packages/gatsby-source-drupal/CHANGELOG.md
@@ -3,6 +3,40 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.0.21](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-source-drupal/compare/gatsby-source-drupal@3.0.20...gatsby-source-drupal@3.0.21) (2019-01-28)
+
+**Note:** Version bump only for package gatsby-source-drupal
+
+<a name="3.0.20"></a>
+
+## [3.0.20](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-source-drupal/compare/gatsby-source-drupal@3.0.19...gatsby-source-drupal@3.0.20) (2019-01-23)
+
+**Note:** Version bump only for package gatsby-source-drupal
+
+<a name="3.0.19"></a>
+
+## [3.0.19](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-source-drupal/compare/gatsby-source-drupal@3.0.18...gatsby-source-drupal@3.0.19) (2019-01-23)
+
+**Note:** Version bump only for package gatsby-source-drupal
+
+<a name="3.0.18"></a>
+
+## [3.0.18](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-source-drupal/compare/gatsby-source-drupal@3.0.16...gatsby-source-drupal@3.0.18) (2019-01-11)
+
+**Note:** Version bump only for package gatsby-source-drupal
+
+<a name="3.0.16"></a>
+
+## [3.0.16](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-source-drupal/compare/gatsby-source-drupal@3.0.15...gatsby-source-drupal@3.0.16) (2019-01-11)
+
+**Note:** Version bump only for package gatsby-source-drupal
+
+<a name="3.0.15"></a>
+
+## [3.0.15](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-source-drupal/compare/gatsby-source-drupal@3.0.14...gatsby-source-drupal@3.0.15) (2019-01-08)
+
+**Note:** Version bump only for package gatsby-source-drupal
+
 <a name="3.0.14"></a>
 
 ## [3.0.14](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-source-drupal/compare/gatsby-source-drupal@3.0.13...gatsby-source-drupal@3.0.14) (2018-12-11)